stt_diary/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai>=1.40.0
pydantic>=2.0.0
python-dotenv  # .env 쓰고 싶으면
